// src/app/shared/generated/model/createToolRequest.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExecutionPolicy } from './executionPolicy';
import { ToolType } from './toolType';
import { AuthMode } from './authMode';


export interface CreateToolRequest { 
    name?: string;
    description?: string;
    type?: ToolType;
    url?: string;
    apiKey?: string;
    executionPolicy?: ExecutionPolicy;
    authMode?: AuthMode;
}